refactor: move anymous element to dedicated modules add tests
